13411 - 하늘에서 정의가 빗발친다!

+++ BaekJoon/13000/13411.cpp
@@ -1,6 +1,6 @@
 // Title : 하늘에서 정의가 빗발친다!
 // Link  : https://www.acmicpc.net/problem/13411 
-// Time  : 40 ms
+// Time  : 36 ms
 // Memory: 3612 KB
 
 #include <bits/stdc++.h>
@@ -25,7 +25,7 @@ int main() {
   stable_sort(arr.begin(), arr.end(), [](const tup& lhs, const tup& rhs) {
     auto [ld, ls, li] = lhs;
     auto [rd, rs, ri] = rhs;
-    return (__int128_t)ld * rs < (__int128_t)rd * ls;
+    return (int64_t)ld * rs < (int64_t)rd * ls;
   });
 
   for (auto [d, s, i] : arr) cout << i << "\n";
